feat: axios refresh token strategy to avoid making requests with old token while we are getting a new one
whenever a 401 happens and it is a token expired error, if it is the first time (!isRefreshing) then we request a refresh token and update the cookies.

For all requests that were attempted meanwhile, they are added to a queue with the configs used.

At the end of the refresh token update process, we retry all the failed requests with their original configuration but using the new token